os: add initial base image for DigitalOcean
I tried to specify the initial configuration.nix, but I couldn't
easily figure it out. It's going to be managed by morph, so I'm not
sweating it. It'd be nice to figure that out eventually. The nixpkgs
revision was entirely arbitrary and happened to be the head of the
release-24.11 branch.