www: add deploy API
I'm not thrilled with this setup, but it'll work for now. I tried my
best to get pam_oidc working, but as it turns out OpenSSH's client
implementation has a hard-coded password length limit of 1024 and
GitHub's OIDC tokens are closer to 4096. A nice improvement would be
to deploy something which can mint short-lived SSH keys, given a valid
OIDC token.