cd: use new uberjar on service restart
The new uberjar (NEXT_PATH) is written by the server when a deployment
is pushed. There's a few seconds of downtime whenever the service
restarts, but I can live with that. At some point, this will need to
run a new and old instance and let Nginx shift traffic to the new one.
That could also help handle the failure case where the new server is
failing to start.